djangopart/requirements.txt
Django==3.2.15
gunicorn==20.0.4
django-rest-framework==0.1.0  
django-cors-headers==3.14.0
Pillow==9.5.0